Rename folders to match components
